toggle active employee works

+++ frontend/src/user/Lists/Edit_List/Confirm/Confirm.js
@@ -5,6 +5,7 @@ import Button from "@material-ui/core/Button";
 //API
 import { list_addItemList } from "../../../../api/api";
 import { list_updateItemList } from "../../../../api/api";
+import { list_deleteItemList } from "../../../../api/api";
 
 //ACTIONS
 import { set_fromStore } from "../../../../actions/actions";
@@ -57,16 +58,17 @@ const Confirm = ({ new_Items, set_isEdit }) => {
 			await list_addItemList(shoppingListID, list_create, token);
 		if (list_update.listItems.length > 0)
 			await list_updateItemList(shoppingListID, list_update, token);
-
-		await fetch(`${url}/${shoppingListID}/items`, {
-			headers: {
-				"Content-Type": "application/json",
-				Authorization: `Bearer ${token}`,
-			},
-			method: "DELETE",
-			body: JSON.stringify(list_delete),
-		});
-		//await list_deleteItemList(shoppingListID, list_delete);
+		if (list_delete.listItems.length > 0) {
+			//await list_deleteItemList(shoppingListID, list_delete, token);
+			await fetch(`${url}/${shoppingListID}/items`, {
+				headers: {
+					"Content-Type": "application/json",
+					Authorization: `Bearer ${token}`,
+				},
+				method: "DELETE",
+				body: JSON.stringify(list_delete),
+			});
+		}
 		dispatch(set_fromStore(false));
 		dispatch(list_getItems(shoppingListID, token));
 		set_isEdit(false);
